+++ monkey/monkey_island/cc/app.py
@@ -169,7 +169,7 @@ def init_api_resources(api):
         "/api/fileUpload/<string:file_type>?load=<string:filename>",
         "/api/fileUpload/<string:file_type>?restore=<string:filename>",
     )
-    api.add_resource(PropagationCredentials, "/api/propagationCredentials")
+    api.add_resource(PropagationCredentials, "/api/propagation-credentials/<string:guid>")
     api.add_resource(RemoteRun, "/api/remote-monkey", "/api/remote-monkey/")
     api.add_resource(VersionUpdate, "/api/version-update", "/api/version-update/")
     api.add_resource(StartedOnIsland, "/api/monkey_control/started_on_island")